+++ src/lib/DeviceInfo.ts
@@ -1,5 +1,3 @@
-import {KnownDevices} from "./KnownDevices";
-
 /**
  * Used to encrypt & decrypt pin. Necessary as long as js-controller can't decrypt for us in array structure.
  * @param key
