--- src/lib/soapclient.d.ts
@@ -1,5 +1,0 @@
-declare module 'soadclient' {
-    function createSoapclient ({user: string, url: string, password: string}) : {
-        login: function;
-    }
-}
